+++ src/shuft/__init__.py
@@ -5,6 +5,7 @@ from pkg_resources import get_distribution
 
 from shuft.shuft import upload
 from shuft.shuft import download
-from shuft.__main__ import cli
+
+from shuft.__main__ import run_command
 
 __version__ = get_distribution(__name__).version